Make common base class for annotation items which render in rect
And avoid a bunch of duplicate code
