+++ python/PyQt6/core/auto_additions/qgsannotationrectitem.py
@@ -1,0 +1,5 @@
+# The following has been generated automatically from src/core/annotations/qgsannotationrectitem.h
+try:
+    QgsAnnotationRectItem.__group__ = ['annotations']
+except NameError:
+    pass
