+++ src/core/annotations/qgsannotationpictureitem.h
@@ -20,7 +20,7 @@
 
 #include "qgis_core.h"
 #include "qgis_sip.h"
-#include "qgsannotationitem.h"
+#include "qgsannotationrectitem.h"
 
 /**
  * \ingroup core
@@ -28,7 +28,7 @@
  *
  * \since QGIS 3.40
  */
-class CORE_EXPORT QgsAnnotationPictureItem : public QgsAnnotationItem
+class CORE_EXPORT QgsAnnotationPictureItem : public QgsAnnotationRectItem
 {
   public:
 
@@ -40,46 +40,14 @@ class CORE_EXPORT QgsAnnotationPictureItem : public QgsAnnotationItem
     ~QgsAnnotationPictureItem() override;
 
     QString type() const override;
-    Qgis::AnnotationItemFlags flags() const override;
-    void render( QgsRenderContext &context, QgsFeedback *feedback ) override;
     bool writeXml( QDomElement &element, QDomDocument &document, const QgsReadWriteContext &context ) const override;
-    QList< QgsAnnotationItemNode > nodesV2( const QgsAnnotationItemEditContext &context ) const override;
-    Qgis::AnnotationItemEditOperationResult applyEditV2( QgsAbstractAnnotationItemEditOperation *operation, const QgsAnnotationItemEditContext &context ) override;
-    QgsAnnotationItemEditOperationTransientResults *transientEditResultsV2( QgsAbstractAnnotationItemEditOperation *operation, const QgsAnnotationItemEditContext &context ) override SIP_FACTORY;
+    bool readXml( const QDomElement &element, const QgsReadWriteContext &context ) override;
+    QgsAnnotationPictureItem *clone() const override SIP_FACTORY;
 
     /**
      * Creates a new polygon annotation item.
      */
     static QgsAnnotationPictureItem *create() SIP_FACTORY;
-
-    bool readXml( const QDomElement &element, const QgsReadWriteContext &context ) override;
-    QgsAnnotationPictureItem *clone() const override SIP_FACTORY;
-    QgsRectangle boundingBox() const override;
-    QgsRectangle boundingBox( QgsRenderContext &context ) const override;
-
-    /**
-     * Returns the bounds of the picture.
-     *
-     * The coordinate reference system for the bounds will be the parent layer's QgsAnnotationLayer::crs().
-     *
-     * When the placementMode() is Qgis::AnnotationPlacementMode::FixedSize then the picture will be placed
-     * at the center of the bounds.
-     *
-     * \see setBounds()
-     */
-    QgsRectangle bounds() const { return mBounds; }
-
-    /**
-     * Sets the \a bounds of the picture.
-     *
-     * The coordinate reference system for the bounds will be the parent layer's QgsAnnotationLayer::crs().
-     *
-     * When the placementMode() is Qgis::AnnotationPlacementMode::FixedSize then the picture will be placed
-     * at the center of the bounds.
-     *
-     * \see bounds()
-     */
-    void setBounds( const QgsRectangle &bounds );
 
     /**
      * Returns the path of the image used to render the item.
@@ -102,56 +70,6 @@ class CORE_EXPORT QgsAnnotationPictureItem : public QgsAnnotationItem
     void setPath( Qgis::PictureFormat format, const QString &path );
 
     /**
-     * Returns the placement mode for the picture.
-     *
-     * \see setPlacementMode()
-     */
-    Qgis::AnnotationPlacementMode placementMode() const;
-
-    /**
-     * Sets the placement \a mode for the picture.
-     *
-     * \see placementMode()
-     */
-    void setPlacementMode( Qgis::AnnotationPlacementMode mode );
-
-    /**
-     * Returns the fixed size to use for the picture, when the placementMode() is Qgis::AnnotationPlacementMode::FixedSize.
-     *
-     * Units are retrieved via fixedSizeUnit()
-     *
-     * \see setFixedSize()
-     * \see fixedSizeUnit()
-     */
-    QSizeF fixedSize() const;
-
-    /**
-     * Sets the fixed \a size to use for the picture, when the placementMode() is Qgis::AnnotationPlacementMode::FixedSize.
-     *
-     * Units are set via setFixedSizeUnit()
-     *
-     * \see fixedSize()
-     * \see setFixedSizeUnit()
-     */
-    void setFixedSize( const QSizeF &size );
-
-    /**
-     * Returns the units to use for fixed picture sizes, when the placementMode() is Qgis::AnnotationPlacementMode::FixedSize.
-     *
-     * \see setFixedSizeUnit()
-     * \see fixedSize()
-     */
-    Qgis::RenderUnit fixedSizeUnit() const;
-
-    /**
-     * Sets the \a unit to use for fixed picture sizes, when the placementMode() is Qgis::AnnotationPlacementMode::FixedSize.
-     *
-     * \see fixedSizeUnit()
-     * \see setFixedSize()
-     */
-    void setFixedSizeUnit( Qgis::RenderUnit unit );
-
-    /**
      * Returns TRUE if the aspect ratio of the picture will be retained.
      *
      * \see setLockAspectRatio()
@@ -165,89 +83,15 @@ class CORE_EXPORT QgsAnnotationPictureItem : public QgsAnnotationItem
      */
     void setLockAspectRatio( bool locked );
 
-    /**
-     * Returns TRUE if the item's background should be rendered.
-     *
-     * \see setBackgroundEnabled()
-     * \see backgroundSymbol()
-     */
-    bool backgroundEnabled() const { return mDrawBackground; }
+  protected:
 
-    /**
-     * Sets whether the item's background should be rendered.
-     *
-     * \see backgroundEnabled()
-     * \see setBackgroundSymbol()
-     */
-    void setBackgroundEnabled( bool enabled ) { mDrawBackground = enabled; }
-
-    /**
-     * Returns the symbol used to render the item's background.
-     *
-     * \see backgroundEnabled()
-     * \see setBackgroundSymbol()
-     */
-    const QgsFillSymbol *backgroundSymbol() const;
-
-    /**
-     * Sets the \a symbol used to render the item's background.
-     *
-     * The item takes ownership of the symbol.
-     *
-     * \see backgroundSymbol()
-     * \see setBackgroundEnabled()
-     */
-    void setBackgroundSymbol( QgsFillSymbol *symbol SIP_TRANSFER );
-
-    /**
-     * Returns TRUE if the item's frame should be rendered.
-     *
-     * \see setFrameEnabled()
-     * \see frameSymbol()
-     */
-    bool frameEnabled() const { return mDrawFrame; }
-
-    /**
-     * Sets whether the item's frame should be rendered.
-     *
-     * \see frameEnabled()
-     * \see setFrameSymbol()
-     */
-    void setFrameEnabled( bool enabled ) { mDrawFrame = enabled; }
-
-    /**
-     * Returns the symbol used to render the item's frame.
-     *
-     * \see frameEnabled()
-     * \see setFrameSymbol()
-     */
-    const QgsFillSymbol *frameSymbol() const;
-
-    /**
-     * Sets the \a symbol used to render the item's frame.
-     *
-     * The item takes ownership of the symbol.
-     *
-     * \see frameSymbol()
-     * \see setBackgroundEnabled()
-     */
-    void setFrameSymbol( QgsFillSymbol *symbol SIP_TRANSFER );
+    void renderInBounds( QgsRenderContext &context, const QRectF &painterBounds, QgsFeedback *feedback ) override;
 
   private:
 
     QString mPath;
     Qgis::PictureFormat mFormat = Qgis::PictureFormat::Unknown;
-    Qgis::AnnotationPlacementMode mPlacementMode = Qgis::AnnotationPlacementMode::SpatialBounds;
-    QgsRectangle mBounds;
-
-    QSizeF mFixedSize;
-    Qgis::RenderUnit mFixedSizeUnit = Qgis::RenderUnit::Millimeters;
-
     bool mLockAspectRatio = true;
-    bool mDrawBackground = false;
-    std::unique_ptr< QgsFillSymbol > mBackgroundSymbol;
-    bool mDrawFrame = false;
-    std::unique_ptr< QgsFillSymbol > mFrameSymbol;
 
 #ifdef SIP_RUN
     QgsAnnotationPictureItem( const QgsAnnotationPictureItem &other );
